mmdet_toolkit/requirements.txt
kymatio==0.3.0 
torchhaarfeatures==0.0.2
yapf==0.40.1
timm>=1.0.17
